Enhancement: Extract RegularExpressionTest
