+++ tests/Constraint/ConstraintTest.php
@@ -1913,56 +1913,6 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintRegularExpression()
-    {
-        $constraint = Assert::matchesRegularExpression('/foo/');
-
-        $this->assertFalse($constraint->evaluate('barbazbar', '', true));
-        $this->assertTrue($constraint->evaluate('barfoobar', '', true));
-        $this->assertEquals('matches PCRE pattern "/foo/"', $constraint->toString());
-        $this->assertCount(1, $constraint);
-
-        try {
-            $constraint->evaluate('barbazbar');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-Failed asserting that 'barbazbar' matches PCRE pattern "/foo/".
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintRegularExpression2()
-    {
-        $constraint = Assert::matchesRegularExpression('/foo/');
-
-        try {
-            $constraint->evaluate('barbazbar', 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-custom message
-Failed asserting that 'barbazbar' matches PCRE pattern "/foo/".
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
     public function testConstraintPCRENotMatch()
     {
         $constraint = Assert::logicalNot(
